fix(chat): polish compact wheel audio followups
* Move compact tool wheel audio preload retry handling into the shared audio helper so App and FullChatSurface do not duplicate timer lifecycle code.

* Clarify rebound playback naming around visual intensity while preserving the disabled rebound-audio behavior.

* Remove no-op rebound playback calls that only returned on the empty rebound source and update the related test description.

* Verified with npm test -- App.test.tsx -t compact tool wheel, npm run typecheck, npm run build, and git diff --check.
